disabled failing qHasFeature_libxml2 to make sure all else working

+++ Library/Sources/Stroika/Foundation/DataExchange/XML/Schema.cpp
@@ -189,7 +189,7 @@ namespace {
                     fSourceComponents.push_back (SourceComponent{.fBLOB = targetNamespaceData, .fNamespace = *targetNamespace});
                 }
                 xmlSchemaParserCtxt* schemaParseContext =
-                    xmlSchemaNewMemParserCtxt (reinterpret_cast<const char*> (targetNamespaceData.data ()), targetNamespaceData.size ());
+                    xmlSchemaNewMemParserCtxt (reinterpret_cast<const char*> (targetNamespaceData.data ()), static_cast<int> (targetNamespaceData.size ()));
                 fCompiledSchema = xmlSchemaParse (schemaParseContext);
                 xmlSchemaFreeParserCtxt (schemaParseContext);
             }
